Moved private KWInequalityType declaration into class .m file.
Because this is a public header, it will be exported to clients of
the library. But the NS_ENUM macro used to declare the
KWInequalityType is only available in the iOS6 and MacOS 10.8 SDK,
while the library would otherwise work fine at least back to the
iOS5 and MacOS 10.6 SDKs (and possibly earlier). And because it is
only used within the class, it is should not be leaked to the
public interface. Moving it to the top of the .m file solves both
issues.