Fix issue where let variables would always be nil
In short, the pointer that the let macro passed into let_ points to the
stack location of the declared variable, whereas when an example node
block capturing a reference to the variable is copied, it now
references the heap location.

So even though the stack variable pointer is dereferenced and assigned
the return value of the block, *the heap version is still nil*.

This is fixed by creating a block, copying it, capturing a reference to
the heap location, and then returning a pointer to it.