Silence deprecation warning for /usr/bin/atos.
`callSiteWithAddress` uses the `atos` command line tool
tool, which outputs the following to stderr when run on OS X 10.9
(13A569) & XCode 5 GM:

```
--
Warning: /usr/bin/atos is moving and will be removed from a future OS X release.
It is now available in the Xcode developer tools to be invoked via: `xcrun atos`
To silence this warning, pass the '-d' command-line flag to this tool.
--
```

This commit silences the warnings.