[XCTEST] Add XCTest subspec.
- Add Kiwi subspec which defines preprocessor macro KIWI_XCTEST.
- KWSpec inherits from XCTestCase or SenTestCase based on value of
  KIWI_XCTEST.

.xctest bundles do not run any tests defined on subclasses of
SenTestCase. Prior to this commit, KWSpec inherits from SenTestCase.
This means that developers creating new projects using Xcode 5, which
includes an .xctest bundle by default, cannot use Kiwi to run tests
on new projects.

Simply changing the base class of KWSpec from SenTestCase to XCTestCase,
on the other hand, causes the opposite problem to occur; .octest bundles
only run tests associated with subclasses of SenTestCase, so the
XCTestCase classes are ignored.

Since there is no way to programmatically determine whether
the test bundle is .xctest or .octest, this commit leaves it up to the
user to specify. Developers using .octest bundles do not need to do
anything; this commit changes the default subspec to be compatible with
SenTestingKit, just as it was before. Developers using .xctest bundles,
however, can specify the "Kiwi/XCTest" subspec in order to have KWSpec
inherit from XCTestCase. This allows the tests to at least be run.
However, XCTestCase requires file names, line numbers, and other
information to be specified for failures. This commit does not include
this information, so individual failures are not reported in the Xcode 5
test navigator.