[XCTEST] Replace KIWI_XCTEST macro with XCT_EXPORT
Instead of defining a new macro, KIWI_XCTEST, the Kiwi/XCTest subspec
now adds an import of the XCTest framework to the Kiwi .pch. That import
results in the definition of the XCT_EXPORT macro, defined by the XCTest
framework. We use this macro to perform conditional logic in the KWSpec
class header and implementation.

This commit also modifies the Kiwi-XCTest target to precompile a
separate header than the Kiwi target, Kiwi-XCTest-Prefix.pch, in order
to mirror the behavior of the subspec.