Add DSL for setup/teardown of entire test suite
Example usage:

    #import <Kiwi/Kiwi.h>

    CONFIG_START

    beforeAllSpecs(^{
        NSLog(@"BEFORE ALL SPECS");
    });

    afterAllSpecs(^{
        NSLog(@"AFTER ALL SPECS");
    });

    CONFIG_END

(Add this code to a file `<tests-dir>/Support/Config.m` and ensure it is
part of your unit test bundle's Compile Sources build phase.)

- - -

Defines a `KWSuiteConfigurationBase` class that provides a
`+defaultConfiguration` singleton instance. The first time this method
is called, we look for a class named `KWSuiteConfiguration` that is a
subclass of `KWSuiteConfigurationBase`, and instantiate it. This class
is declared by using the `CONFIG_START` and `CONFIG_END` macros.

When the default configuration receives the `-setUp` message, it sends
`-configureSuite` to `self`, which executes the code between
`CONFIG_START` and `CONFIG_END`.

The blocks passed to `beforeAllSpecs` and `afterAllSpecs` are then
executed in `-setUp` and `-tearDown`, respectively.