Correctly register KWConfigurationTestObserver as XCTestObserverClass
There is a timing issue when setting the `XCTestObserverClass` default via
`setObject:forKey:` -- when called from `+load`, XCTest's own startup overrides
the value that was set, and when called from `+initialize` it seems to be too
late for XCTest to actually use the value (i.e., it's already created
observers).

Turns out, XCTest actually respects a default for `XCTestObserverClass` set in
`NSRegistrationDomain`. This means that the defaults registration can be moved
into `+load`, ensuring it happens before XCTest sets up observers.