Fix crash caused by over-released object when using receive:withArguments:
This crash has manifested itself variously as a SIGABRT in an unexpected
location, and an EXC_BAD_ACCESS when leaving an @autoreleasepool { }
scope. It seems to have been introduced when converting KWMessagePattern
to ARC, causing message patterns with argument filters to be over-eager
in releasing their arguments.

The crash is resolved by replacing strong references to message pattern
argument filters with autoreleasing strong references.

(You can reproduce the crash consistently and clearly by reverting the
offending line in KWMessagePattern, then wrapping the new functional
test in an @autoreleasepool { } scope.)