Add (failing) functional test describing use of future strings in expectations
When futures are involved, it seems that -isEqual: is not necessarily
commutative. For example, these two aren't guaranteed to be equivalent:

    [[futureObject should] equal:actualObject];
    [[actualObject should] equal:futureObject];

In the case of string comparison, it is the second form that fails. An
MAFuture is very lightweight proxy for the resolved future value --
almost all messages the future receives are forwarded to the real object
it represents. However when a future string is on the right-hand side of
-isEqual:, the comparison fails. My best guess is that the
implementation of -isEqualToString: doesn't send any messages to the
right operand, and accesses its internal storage directly.